Add hilbertBitRect to js prototype
This function shows the unique rectangle to all points in a hilbert curve whose binary "id" starts with a certain prefix.
For example when I use this in the steno viz it'll allow you to see where all the "SKP~" entries live.

Furthermore I'm not sure if finding the bounding box over 4 points is ideal, but the code was a lot shorter than my first approach, which was to keep track of the orientation of a point on the hilbert curve. This resulted in essentially duplicate code from the hilbert(t) function, and the return statement had 5 different cases. I'm not sure if it's faster, but when I write the C++ version I may re-write it and test both ways.